+++ src/features/pmicControl/npm/pmic1300Device.test.ts
@@ -556,7 +556,7 @@ describe('PMIC 1300', () => {
                 );
             });
 
-            test('Request getDefaultBatteryModels success', async () => {
+            test('Request getHardcodedBatteryModels success', async () => {
                 mockEnqueueRequest.mockImplementationOnce(
                     (
                         _command: string,
@@ -569,11 +569,13 @@ describe('PMIC 1300', () => {
                         callbacks?.onSuccess(
                             `Currently active battery model:
                             name="LP803448",T={5.00 C,25.00 C,45.00 C},Q={1413.40 mAh,1518.28 mAh,1500.11 mAh}
-                    Default battery models:
+                    Hardcoded battery models:
                             name="LP803448",T={5.00 C,25.00 C,45.00 C},Q={1413.40 mAh,1518.28 mAh,1500.11 mAh}
                             name="LP502540",T={25.00 C},Q={563.08 mAh}
-                    Battery model stored in database:
-                            name="LP803448",T={5.00 C,25.00 C,45.00 C},Q={1413.40 mAh,1518.28 mAh,1500.11 mAh}`,
+                    Battery models stored in database:
+                            Slot 0: Empty
+                            Slot 1: Empty
+                            Slot 2: Empty`,
                             'fuel_gauge model list'
                         );
                         return Promise.resolve();
@@ -581,7 +583,7 @@ describe('PMIC 1300', () => {
                 );
 
                 await expect(
-                    pmic.getDefaultBatteryModels()
+                    pmic.getHardcodedBatteryModels()
                 ).resolves.toStrictEqual([
                     {
                         name: 'LP803448',
@@ -1405,21 +1407,6 @@ describe('PMIC 1300', () => {
                     );
                 }
             );
-
-            test('storeBattery', async () => {
-                await pmic.storeBattery();
-
-                expect(mockEnqueueRequest).toBeCalledTimes(1);
-                expect(mockEnqueueRequest).toBeCalledWith(
-                    'fuel_gauge model store',
-                    expect.anything(),
-                    undefined,
-                    true
-                );
-
-                // Updates should only be emitted when we get response
-                expect(mockOnActiveBatteryModelUpdate).toBeCalledTimes(0);
-            });
         });
 
         describe('Setters and effects state - error', () => {
@@ -3169,7 +3156,7 @@ describe('PMIC 1300', () => {
             expect(mockOnFuelGaugeUpdate).toBeCalledWith(enabled);
         });
 
-        test.each(['get', 'set LP803448'])('fuel_gauge model %p', append => {
+        test.each(['get', 'set "LP803448"'])('fuel_gauge model %p', append => {
             const command = `fuel_gauge model ${append}`;
             const callback =
                 eventHandlers.mockRegisterCommandCallbackHandler(command);
@@ -3233,36 +3220,42 @@ describe('PMIC 1300', () => {
 
             const response = `Currently active battery model:
             name="LP803448",T={5.00 C,25.00 C,45.00 C},Q={1413.40 mAh,1518.28 mAh,1500.11 mAh}
-    Default battery models:
+    Hardcoded battery models:
             name="LP302535",T={5.00 C,25.00 C,45.00 C},Q={273.95 mAh,272.80 mAh,269.23 mAh}
             name="LP353035",T={5.00 C,25.00 C,45.00 C},Q={406.15 mAh,422.98 mAh,420.56 mAh}
             name="LP301226",T={5.00 C,25.00 C,45.00 C},Q={68.99 mAh,70.43 mAh,65.50 mAh}
             name="LP803448",T={5.00 C,25.00 C,45.00 C},Q={1413.40 mAh,1518.28 mAh,1500.11 mAh}
             name="LP502540",T={25.00 C},Q={563.08 mAh}
             name="LP503030",T={25.00 C},Q={495.98 mAh}
-    Battery model stored in database:
-            name="LP803448",T={5.00 C,25.00 C,45.00 C},Q={1413.40 mAh,1518.28 mAh,1500.11 mAh}`;
+    Battery models stored in database:
+            Slot 0: Empty
+            Slot 1: name="LP803448",T={5.00 C,25.00 C,45.00 C},Q={1413.40 mAh,1518.28 mAh,1500.11 mAh}
+            Slot 2: Empty`;
 
             callback?.onSuccess(response, command);
 
             expect(mockOnStoredBatteryModelUpdate).toBeCalledTimes(1);
-            expect(mockOnStoredBatteryModelUpdate).toBeCalledWith({
-                name: 'LP803448',
-                characterizations: [
-                    {
-                        temperature: 45,
-                        capacity: 1500.11,
-                    },
-                    {
-                        temperature: 25,
-                        capacity: 1518.28,
-                    },
-                    {
-                        temperature: 5,
-                        capacity: 1413.4,
-                    },
-                ],
-            } as BatteryModel);
+            expect(mockOnStoredBatteryModelUpdate).toBeCalledWith([
+                null,
+                {
+                    name: 'LP803448',
+                    characterizations: [
+                        {
+                            temperature: 45,
+                            capacity: 1500.11,
+                        },
+                        {
+                            temperature: 25,
+                            capacity: 1518.28,
+                        },
+                        {
+                            temperature: 5,
+                            capacity: 1413.4,
+                        },
+                    ],
+                },
+                null,
+            ] as (BatteryModel | null)[]);
         });
 
         test('fuel_gauge model list no stored battery', () => {
@@ -3272,17 +3265,25 @@ describe('PMIC 1300', () => {
 
             const response = `Currently active battery model:
                 name="LP803448",T={5.00 C,25.00 C,45.00 C},Q={1413.40 mAh,1518.28 mAh,1500.11 mAh}
-        Default battery models:
+            Hardcoded battery models:
                 name="LP302535",T={5.00 C,25.00 C,45.00 C},Q={273.95 mAh,272.80 mAh,269.23 mAh}
                 name="LP353035",T={5.00 C,25.00 C,45.00 C},Q={406.15 mAh,422.98 mAh,420.56 mAh}
                 name="LP301226",T={5.00 C,25.00 C,45.00 C},Q={68.99 mAh,70.43 mAh,65.50 mAh}
                 name="LP803448",T={5.00 C,25.00 C,45.00 C},Q={1413.40 mAh,1518.28 mAh,1500.11 mAh}
                 name="LP502540",T={25.00 C},Q={563.08 mAh}
-                name="LP503030",T={25.00 C},Q={495.98 mAh}`;
+                name="LP503030",T={25.00 C},Q={495.98 mAh}
+            Battery models stored in database:
+                Slot 0: Empty
+                Slot 1: Empty
+                Slot 2: Empty`;
             callback?.onSuccess(response, command);
 
             expect(mockOnStoredBatteryModelUpdate).toBeCalledTimes(1);
-            expect(mockOnStoredBatteryModelUpdate).toBeCalledWith(undefined);
+            expect(mockOnStoredBatteryModelUpdate).toBeCalledWith([
+                null,
+                null,
+                null,
+            ]);
         });
 
         test.each([true, false])('npmx vbusin vbus status get %p', value => {